Repair low level "hiding" items
The cursor was still on the hidden items because the text property
'front-sticky was not set. I do not understand why this is a problem,
but now it works. Hiding an item means to "fold" it away, leaving all
other items at their positions even if they go up visually one line.