Speed up insertion of longer sequences.
The problem is that `lister-add-marker` takes long time since it has
to sort the whole list each time a marker is added. So we add an
option to insert an item or a sequence without updating the marker
list. Thus we can insert a long sequence and update the marker list
only once, instead of updating it each time an item is inserted.