Further simplified filtering to prevent mistakes
There had been several functions which had the effect of clearing the
list by filtering all items, yet only one of them SHOULD be used
because it alone handles the buffer local filter function, which is
important for insertion and for the macro `lister-with-locked-cursor'.
We sometimes need to know if a filter is active, period. So we cut off
all these misleading alternatives, forcing the user to go the way he
or she should go.