lister-add-item-marker: use cl-sort
This simplifies the code, but the result is still not satisfying. The
problem remains that in whatever way this function is currently
implemented, it always requires to traverse the list to determine the
place to insert at. So I opt for a clean code here, instead of trying
to optimize the hell out of it.