Wrapped lists: allow input lists to begin with a list
Previous version of "lister--wrap-list" used to return wrong results
when fed with a list whose car already was a list. That made sense
insofar it was assumed that we only "wrap" sublists if there is a
corresponding heading node; but that made sorting useless for
arbitrary regions, where such an input might be actually the case. So
this new version also works correctly when feeding lists which do not
begin with a non-list item.

Rewrite lister--wrap-list; change lister--reorder-wrapped-list; add tests.