Implement sorting multiple times
This commit extends the sorting command to accept a list of
predicates (comparators) instead of only one single function. In that
new case, the list will be sorted by using the first predicate first,
then passing the result to the sorting algorithm using the second
predicate, etc.

This is useful for sorting against multiple criteria; e.g. sort first
according to the last name, then according to the first name.