Fix buggy refresh of  visible top nodes of invisible lists
Closes issue #12 opened for Delve. If a sublist was invisible and its
top item had been refreshed, the new item was inserted under the
invisible overlay and thus remained itself invisible. This has been
fixed now by changing the insertion behavior of the overlay's start
and end marker.

This fix still leaves an undesired behavior: In some cases, the
ellipsis representing hidden items is inserted on a separate line, in
other, preferred, cases as the final end of the last visible item.